# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "School Hostel Enterprise API",
    "description": "Enterprise-grade API for school academics and residential hostel operations. It includes RBAC, students, attendance, fee collection, hostel occupancy, mess, visitors, leave approvals, complaints, dashboards, and audit logging.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/hostel/buildings": {
      "get": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "List Buildings",
        "operationId": "list_buildings_api_v1_hostel_buildings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BuildingRead"
                  },
                  "type": "array",
                  "title": "Response List Buildings Api V1 Hostel Buildings Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "Create Building",
        "operationId": "create_building_api_v1_hostel_buildings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildingRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/hostel/rooms": {
      "post": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "Create Room",
        "operationId": "create_room_api_v1_hostel_rooms_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "List Rooms",
        "operationId": "list_rooms_api_v1_hostel_rooms_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "building_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Building Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomRead"
                  },
                  "title": "Response List Rooms Api V1 Hostel Rooms Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostel/beds": {
      "get": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "List Beds",
        "operationId": "list_beds_api_v1_hostel_beds_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "room_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Room Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/BedStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BedRead"
                  },
                  "title": "Response List Beds Api V1 Hostel Beds Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostel/allocations": {
      "post": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "Allocate Bed",
        "operationId": "allocate_bed_api_v1_hostel_allocations_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AllocationCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "List Allocations",
        "operationId": "list_allocations_api_v1_hostel_allocations_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Student Id"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/AllocationStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AllocationRead"
                  },
                  "title": "Response List Allocations Api V1 Hostel Allocations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostel/allocations/{allocation_id}/checkout": {
      "post": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "Checkout Student",
        "operationId": "checkout_student_api_v1_hostel_allocations__allocation_id__checkout_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "allocation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Allocation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AllocationRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/hostel/occupancy": {
      "get": {
        "tags": [
          "Hostel Management"
        ],
        "summary": "Occupancy Dashboard",
        "operationId": "occupancy_dashboard_api_v1_hostel_occupancy_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AllocationCreate": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "bed_id": {
            "type": "integer",
            "title": "Bed Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "deposit_amount": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Deposit Amount",
            "default": "0.00"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "bed_id",
          "start_date"
        ],
        "title": "AllocationCreate"
      },
      "AllocationRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "bed_id": {
            "type": "integer",
            "title": "Bed Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "deposit_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Deposit Amount"
          },
          "status": {
            "$ref": "#/components/schemas/AllocationStatus"
          },
          "allocated_by": {
            "type": "integer",
            "title": "Allocated By"
          },
          "checkout_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "bed_id",
          "start_date",
          "end_date",
          "deposit_amount",
          "status",
          "allocated_by",
          "checkout_notes"
        ],
        "title": "AllocationRead"
      },
      "AllocationStatus": {
        "type": "string",
        "enum": [
          "active",
          "completed",
          "cancelled"
        ],
        "title": "AllocationStatus"
      },
...
      "BedRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "room_id": {
            "type": "integer",
            "title": "Room Id"
          },
          "bed_no": {
            "type": "string",
            "title": "Bed No"
          },
          "status": {
            "$ref": "#/components/schemas/BedStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "room_id",
          "bed_no",
          "status"
        ],
        "title": "BedRead"
      },
      "BedStatus": {
        "type": "string",
        "enum": [
          "available",
          "occupied",
          "reserved",
          "maintenance"
        ],
        "title": "BedStatus"
      },
...
      "BuildingCreate": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender_type": {
            "$ref": "#/components/schemas/Gender"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "warden_staff_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warden Staff Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "gender_type"
        ],
        "title": "BuildingCreate"
      },
      "BuildingRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "gender_type": {
            "$ref": "#/components/schemas/Gender"
          },
          "address": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "warden_staff_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warden Staff Id"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "gender_type",
          "address",
          "warden_staff_id",
          "is_active"
        ],
        "title": "BuildingRead"
      },
      "CheckoutRequest": {
        "properties": {
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "checkout_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Notes"
          }
        },
        "type": "object",
        "required": [
          "end_date"
        ],
        "title": "CheckoutRequest"
      },
...
      "Gender": {
        "type": "string",
        "enum": [
          "male",
          "female",
          "other"
        ],
        "title": "Gender"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoomCreate": {
        "properties": {
          "building_id": {
            "type": "integer",
            "title": "Building Id"
          },
          "room_no": {
            "type": "string",
            "title": "Room No"
          },
          "floor": {
            "type": "integer",
            "title": "Floor",
            "default": 0
          },
          "capacity": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Capacity"
          },
          "monthly_fee": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Monthly Fee"
          },
          "status": {
            "$ref": "#/components/schemas/RoomStatus",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "building_id",
          "room_no",
          "capacity",
          "monthly_fee"
        ],
        "title": "RoomCreate"
      },
      "RoomRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "building_id": {
            "type": "integer",
            "title": "Building Id"
          },
          "room_no": {
            "type": "string",
            "title": "Room No"
          },
          "floor": {
            "type": "integer",
            "title": "Floor"
          },
          "capacity": {
            "type": "integer",
            "title": "Capacity"
          },
          "monthly_fee": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Monthly Fee"
          },
          "status": {
            "$ref": "#/components/schemas/RoomStatus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "building_id",
          "room_no",
          "floor",
          "capacity",
          "monthly_fee",
          "status"
        ],
        "title": "RoomRead"
      },
      "RoomStatus": {
        "type": "string",
        "enum": [
          "active",
          "maintenance",
          "closed"
        ],
        "title": "RoomStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}